Fixes, Spell, Underlayers, Placeholders, VFX, New Fonts, New Icons

+++ Assets/_Scripts/Combat/UnitStats.cs
@@ -63,7 +63,7 @@ public struct UnitStats : INotifyPropertyChanged
     public void SetSnapshot(IUnitStatsProvider statsProvider)
     {
         Attack = (int)statsProvider.GetStats(UnitParameterType.Attack);
-        AttackSpeed = (int)statsProvider.GetStats(UnitParameterType.AttackSpeed);
+        AttackSpeed = statsProvider.GetStats(UnitParameterType.AttackSpeed);
         MaxHealth = (int)statsProvider.GetStats(UnitParameterType.MaxHealth);
         UpgradeCost = (int)statsProvider.GetStats(UnitParameterType.UpgradeCost);
     }
